+++ lib/Login/otp.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:side_menu/Reusable/app_input_text.dart';
 import 'package:side_menu/Reusable/app_input_textfield.dart';
+import 'package:side_menu/Reusable/button_component.dart';
 
-import 'Reusable/button_component.dart';
 
 class otp extends StatefulWidget {
   const otp({super.key});
