+++ lib/PrescriptionView/prescription_view.dart
@@ -7,7 +7,6 @@ import 'package:flutter/src/widgets/framework.dart';
 import 'package:side_menu/Reusable/app_input_textfield.dart';
 import 'package:side_menu/Reusable/date_picker.dart';
 
-import 'Reusable/button_component.dart';
 
 class viewPrescription extends StatefulWidget {
   const viewPrescription({super.key});
