+++ lib/Reusable/alert_for_medicineData.dart
@@ -19,6 +19,8 @@ import 'package:universal_io/io.dart';
 import '../modelClasses/medicine_data_model.dart';
 import '../modelClasses/medicine_list_provider.dart';
 import 'alertInputTextfield.dart';
+import 'alert_date_picker.dart';
+import 'alert_textformfield.dart';
 
 class AppShowAlertMedicineData extends StatelessWidget {
   const AppShowAlertMedicineData({
@@ -50,31 +52,91 @@ class AppShowAlertMedicineData extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final _formkey1 = GlobalKey<FormState>();
+    final _formkey2 = GlobalKey<FormState>();
     final studentsStateProvider = Provider.of<MedicineListProvider>(context);
     MedicinenameController.text = "";
     ExpiryDateController.text = "";
     String? MedicinenameControllerText;
-    return new Center(child: (() {
-      print('OS: ${Platform.operatingSystem}');
-      if (defaultTargetPlatform == TargetPlatform.android) {
-        return new AlertDialog(
-            backgroundColor: Colors.blueAccent,
-            title: Text(message),
-            content: Column(
-              children: [
-                AppInputTextfield(
+    return SingleChildScrollView(
+      child: new Center(child: (() {
+        print('OS: ${Platform.operatingSystem}');
+        if (defaultTargetPlatform == TargetPlatform.android) {
+          return new AlertDialog(
+              insetPadding: EdgeInsets.symmetric(
+                horizontal: 50.0,
+                vertical: 150.0,
+              ),
+              backgroundColor: Colors.white,
+              title: Text(message),
+              content: Column(
+                children: [
+                  AlertTextFormField(
+                      hintText: hintText,
+                      nameController: MedicinenameController,
+                      onChanged: (value) {
+                        MedicinenameControllerText = value;
+                      },
+                      errorMessage: errorMessage,
+                      input_type: input_type,
+                      obsecuretext: obsecuretext,
+                      node: node,
+                      action: action,
+                      onEditingComplete: onEditingComplete,
+                      globalKey: _formkey1),
+                  AlertdatePickerComponent(
+                      hintText: hintText1,
+                      nameController: ExpiryDateController,
+                      errorMessage: errorMessage1,
+                      obsecuretext: obsecuretext,
+                      node: node,
+                      action: action,
+                      onEditingComplete: onEditingComplete,
+                      globalKey: _formkey2)
+                ],
+              ),
+              actions: [
+                TextButton(
+                    onPressed: (() {
+                      Navigator.pop(context);
+                    }),
+                    child: AppInputText(
+                        text: 'Cancel',
+                        colors: Colors.blue,
+                        size: 20,
+                        weight: FontWeight.bold)),
+                TextButton(
+                    onPressed: (() {
+                      if (_formkey1.currentState!.validate() &&
+                          _formkey2.currentState!.validate()) {
+                        studentsStateProvider.addMedicineData(medicineDataModel(
+                            medicineName: MedicinenameController.text,
+                            ExpiryDate: ExpiryDateController.text));
+                        print("data added successfully" +
+                            studentsStateProvider.Medicines.length.toString());
+                        Navigator.pop(context);
+                      }
+                    }),
+                    child: AppInputText(
+                        text: 'OK',
+                        colors: Colors.blue,
+                        size: 20,
+                        weight: FontWeight.bold))
+              ]);
+        } else {
+          return new CupertinoAlertDialog(
+              title: Text(message),
+              content: Column(children: [
+                AlertTextFormField(
                     hintText: hintText,
                     nameController: MedicinenameController,
-                    onChanged: (value) {
-                      MedicinenameControllerText = value;
-                    },
                     errorMessage: errorMessage,
                     input_type: input_type,
                     obsecuretext: obsecuretext,
                     node: node,
                     action: action,
                     onEditingComplete: onEditingComplete),
-                datePickerComponent(
+                AlertdatePickerComponent(
                     hintText: hintText1,
                     nameController: ExpiryDateController,
                     errorMessage: errorMessage1,
@@ -82,76 +144,35 @@ class AppShowAlertMedicineData extends StatelessWidget {
                     node: node,
                     action: action,
                     onEditingComplete: onEditingComplete)
-              ],
-            ),
-            actions: [
-              TextButton(
-                  onPressed: (() {
-                    Navigator.pop(context);
-                  }),
-                  child: AppInputText(
-                      text: 'cancel',
-                      colors: Colors.white,
-                      size: 20,
-                      weight: FontWeight.bold)),
-              TextButton(
-                  onPressed: (() {
-                    studentsStateProvider.addMedicineData(medicineDataModel(
-                        medicineName: MedicinenameController.text,
-                        ExpiryDate: ExpiryDateController.text));
-                    print("data added successfully" +
-                        studentsStateProvider.Medicines.length.toString());
-                    Navigator.pop(context);
-                  }),
-                  child: AppInputText(
-                      text: 'OK',
-                      colors: Colors.white,
-                      size: 20,
-                      weight: FontWeight.bold))
-            ]);
-      } else {
-        return new CupertinoAlertDialog(
-            title: Text(message),
-            content: Column(children: [
-              AppInputTextfield(
-                  hintText: hintText,
-                  nameController: MedicinenameController,
-                  errorMessage: errorMessage,
-                  input_type: input_type,
-                  obsecuretext: obsecuretext,
-                  node: node,
-                  action: action,
-                  onEditingComplete: onEditingComplete),
-              datePickerComponent(
-                  hintText: hintText1,
-                  nameController: ExpiryDateController,
-                  errorMessage: errorMessage1,
-                  obsecuretext: obsecuretext,
-                  node: node,
-                  action: action,
-                  onEditingComplete: onEditingComplete)
-            ]),
-            actions: [
-              TextButton(
-                  onPressed: (() {
-                    Navigator.pop(context);
-                  }),
-                  child: AppInputText(
-                      text: 'Cancel',
-                      colors: Colors.white,
-                      size: 20,
-                      weight: FontWeight.bold)),
-              TextButton(
-                  onPressed: (() {
-                    Navigator.pop(context);
-                  }),
-                  child: AppInputText(
-                      text: 'OK',
-                      colors: Colors.white,
-                      size: 20,
-                      weight: FontWeight.bold))
-            ]);
-      }
-    })());
+              ]),
+              actions: [
+                TextButton(
+                    onPressed: (() {
+                      Navigator.pop(context);
+                    }),
+                    child: AppInputText(
+                        text: 'Cancel',
+                        colors: Colors.blue,
+                        size: 20,
+                        weight: FontWeight.bold)),
+                TextButton(
+                    onPressed: (() {
+                      studentsStateProvider.addMedicineData(medicineDataModel(
+                          medicineName: MedicinenameController.text,
+                          ExpiryDate: ExpiryDateController.text));
+                      print("data added successfully" +
+                          studentsStateProvider.Medicines.length.toString());
+                      Navigator.pop(context);
+                      Navigator.pop(context);
+                    }),
+                    child: AppInputText(
+                        text: 'OK',
+                        colors: Colors.blue,
+                        size: 20,
+                        weight: FontWeight.bold))
+              ]);
+        }
+      })()),
+    );
   }
 }
