+++ lib/Routes/App_routes.dart
@@ -1,11 +1,3 @@
-import 'package:side_menu/add_prescription.dart';
-
-import 'package:side_menu/dashboard_gridview.dart';
-import 'package:side_menu/family_list.dart';
-import 'package:side_menu/mpin.dart';
-import 'package:side_menu/prescription_list.dart';
-import 'package:side_menu/prescription_view.dart';
-import 'package:side_menu/visit_alerts.dart';
 
 class AppRoutes {
   AppRoutes._();
