+++ pubspec.yaml
@@ -48,6 +48,7 @@ dependencies:
   provider: ^6.0.4
   flutter_svg: ^1.1.6
   fluttertoast: ^8.1.1
+  flutter_group_button: ^2.0.1
 
 flutter_icons:
   image_path: "assets/appLogo.png"
